requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Document Processing
PyMuPDF==1.23.8
python-pptx==0.6.23
Pillow==10.1.0
pytesseract==0.3.10

# AI and Machine Learning
google-generativeai==0.3.2
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.4

# PDF Generation
fpdf2==2.7.6

# Data Processing
pandas==2.1.3

# Utilities
pydantic==2.5.0